requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docx2txt==0.8
Flask==2.3.2
gensim==4.3.1
gitdb==4.0.10
GitPython==3.1.31
milvus==2.2.10
mysql-connector-python==8.0.33
nltk==3.8.1
oauthlib==3.2.0
openai==0.27.8
opencv-python==4.7.0.72
pdf2image==1.16.3
pdfminer.six==20191110
pymilvus==2.2.13
PyMuPDF==1.22.5
PyPDF2==3.0.1
pytesseract==0.3.10
python-docx==0.8.11
python-dotenv==1.0.0
regex==2023.6.3
SQLAlchemy==2.0.17
streamlit==1.24.1
zipp==1.0.0
